Turn uint32 into uintptr_t in WrapNativeTexture.
This fixes a compile error on XboxOne.

+++ media/base/video_frame.cc
@@ -62,7 +62,7 @@ bool VideoFrame::IsValidConfig(VideoFrame::Format format,
 
 // static
 scoped_refptr<VideoFrame> VideoFrame::WrapNativeTexture(
-    uint32 texture_id,
+    uintptr_t texture_id,
     uint32 texture_target,
     const gfx::Size& coded_size,
     const gfx::Rect& visible_rect,
